LPS-99657 Rename 'Action' -> 'StrutsAction' for consistency
